+++ mmrazor/models/utils/build_functions/__init__.py
@@ -1,8 +1,10 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .prune_deploy_sub_model import PruneDeploySubModel
 from .prune_sub_model import PruneSubModel
+from .sub_model import sub_model
 
 __all__ = [
     'PruneDeploySubModel',
     'PruneSubModel',
+    'sub_model',
 ]
