+++ mmrazor/registry/__init__.py
@@ -3,13 +3,25 @@ from .registry import (DATA_SAMPLERS, DATASETS, HOOKS, LOOPS, METRICS,
                        MODEL_WRAPPERS, MODELS, OPTIM_WRAPPER_CONSTRUCTORS,
                        OPTIM_WRAPPERS, OPTIMIZERS, PARAM_SCHEDULERS,
                        RUNNER_CONSTRUCTORS, RUNNERS, TASK_UTILS, TRANSFORMS,
-                       VISBACKENDS, VISUALIZERS, WEIGHT_INITIALIZERS,
-                       sub_model)
+                       VISBACKENDS, VISUALIZERS, WEIGHT_INITIALIZERS)
 
 __all__ = [
-    'RUNNERS', 'RUNNER_CONSTRUCTORS', 'HOOKS', 'DATASETS', 'DATA_SAMPLERS',
-    'TRANSFORMS', 'MODELS', 'WEIGHT_INITIALIZERS', 'OPTIMIZERS',
-    'OPTIM_WRAPPERS', 'OPTIM_WRAPPER_CONSTRUCTORS', 'TASK_UTILS',
-    'PARAM_SCHEDULERS', 'METRICS', 'MODEL_WRAPPERS', 'LOOPS', 'VISBACKENDS',
-    'VISUALIZERS', 'sub_model'
+    'RUNNERS',
+    'RUNNER_CONSTRUCTORS',
+    'HOOKS',
+    'DATASETS',
+    'DATA_SAMPLERS',
+    'TRANSFORMS',
+    'MODELS',
+    'WEIGHT_INITIALIZERS',
+    'OPTIMIZERS',
+    'OPTIM_WRAPPERS',
+    'OPTIM_WRAPPER_CONSTRUCTORS',
+    'TASK_UTILS',
+    'PARAM_SCHEDULERS',
+    'METRICS',
+    'MODEL_WRAPPERS',
+    'LOOPS',
+    'VISBACKENDS',
+    'VISUALIZERS',
 ]
